+++ components/modal/src/modal-title/modal-title.js
@@ -3,7 +3,10 @@ import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-export const ModalTitle = ({ children, dataTest }) => (
+export const ModalTitle = ({
+    children,
+    dataTest = 'dhis2-uicore-modaltitle',
+}) => (
     <h1 className={cx('title')} data-test={dataTest}>
         {children}
 
@@ -20,10 +23,6 @@ export const ModalTitle = ({ children, dataTest }) => (
         `}</style>
     </h1>
 )
-
-ModalTitle.defaultProps = {
-    dataTest: 'dhis2-uicore-modaltitle',
-}
 
 ModalTitle.propTypes = {
     children: PropTypes.node,