+++ components/popper/src/popper.js
@@ -15,15 +15,17 @@ const flipPlacement = (placement) => {
     return placement
 }
 
+// Stable object reference
+const staticArray = []
 const Popper = ({
     children,
     className,
-    dataTest,
-    modifiers,
+    dataTest = 'dhis2-uicore-popper',
+    modifiers = staticArray,
     observePopperResize,
     observeReferenceResize,
     onFirstUpdate,
-    placement,
+    placement = 'auto',
     reference,
     strategy,
 }) => {
@@ -67,12 +69,6 @@ const Popper = ({
             {children}
         </div>
     )
-}
-
-Popper.defaultProps = {
-    dataTest: 'dhis2-uicore-popper',
-    modifiers: [],
-    placement: 'auto',
 }
 
 // Prop names follow the names here: https://popper.js.org/docs/v2/constructors/